+++ servers/mcp-neo4j-data-modeling/CHANGELOG.md
@@ -9,6 +9,9 @@
 * Change default transport to `stdio` in Dockerfile
 
 ### Added
+* Add security middleware (CORS and TrustedHost) for HTTP and SSE transports
+* Add CLI  support for `--allow-origins` and `--allowed-hosts` configuration
+* Add environment variable for `NEO4J_MCP_SERVER_ALLOW_ORIGINS` and `NEO4J_MCP_SERVER_ALLOWED_HOSTS` configuration
 
 ## v0.4.0